requirements.txt
Flask
gunicorn
Flask-Cors
flask-sqlalchemy
Flask-MySQLdb
flask_jwt_simple
Flask-Mail
secure-smtplib
skpy
requests
google-api-python-client
google-auth-httplib2
google-auth-oauthlib